+++ docs/make.jl
@@ -40,6 +40,7 @@ makedocs(;
             "Export results" => "how-to/export-results.md",
             "Customize colors" => "how-to/customize-colors.md",
             "Customize icons" => "how-to/customize-icons.md",
+            "Customize descriptive_names" => "how-to/customize-descriptive_names.md",
         ],
         "Library" => Any[
             "Public" => "library/public.md",